use preferred title mode everywhere a title is shown
