+++ app/src/main/java/io/github/shadow578/tenshi/adapter/SeasonalAnimeAdapter.java
@@ -16,7 +16,9 @@ import io.github.shadow578.tenshi.databinding.RecyclerAnimeSmallBinding;
 import io.github.shadow578.tenshi.extensionslib.lang.BiConsumer;
 import io.github.shadow578.tenshi.extensionslib.lang.Consumer;
 import io.github.shadow578.tenshi.mal.model.AnimeRankingItem;
+import io.github.shadow578.tenshi.mal.model.type.TitleDisplayMode;
 import io.github.shadow578.tenshi.util.GlideHelper;
+import io.github.shadow578.tenshi.util.TenshiPrefs;
 
 import static io.github.shadow578.tenshi.extensionslib.lang.LanguageUtil.*;
 
@@ -75,7 +77,8 @@ public class SeasonalAnimeAdapter extends RecyclerView.Adapter<SeasonalAnimeAdap
                 -> GlideHelper.glide(ctx, p.mediumUrl).into(b.animeMainPoster));
 
         // title
-        b.animeTitle.setText(elvisEmpty(anime.anime.title, unknown));
+        final TitleDisplayMode titleMode = TenshiPrefs.getEnum(TenshiPrefs.Key.TitleDisplayMode, TitleDisplayMode.class, TitleDisplayMode.Canonical);
+        b.animeTitle.setText(elvisEmpty(anime.anime.getDisplayTitle(titleMode), unknown));
 
         // setup onclick listener
         h.itemView.setOnClickListener(view -> onClickListener.invoke(b.animeMainPoster, anime));
