+++ app/src/main/java/io/github/shadow578/tenshi/mal/model/type/TitleDisplayMode.java
@@ -1,0 +1,21 @@
+package io.github.shadow578.tenshi.mal.model.type;
+
+/**
+ * modes a anime title can be displayed in
+ */
+public enum TitleDisplayMode {
+    /**
+     * prefer the canonical title
+     */
+    Canonical,
+
+    /**
+     * prefer the english title
+     */
+    English,
+
+    /**
+     * prefer the japanese title
+     */
+    Japanese
+}
